Only add index entries instead of modifying entity
This allows the datastore to avoid the step of decoding and
re-encoding entities in order to strip reserved properties.
